feat: set userAgent in RNSession
