+++ packages/turbo/android/src/main/java/com/reactnativeturbowebview/RNSession.kt
@@ -1,6 +1,7 @@
 package com.reactnativeturbowebview
 
 import android.webkit.JavascriptInterface
+import android.webkit.WebSettings
 import android.webkit.WebView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.Lifecycle
@@ -24,6 +25,7 @@ import org.json.JSONObject
 class RNSession(
   private val reactContext: ReactApplicationContext,
   private val sessionHandle: String,
+  private val applicationNameForUserAgent: String?
 ) : SessionCallbackAdapter {
 
   var visitableView: SessionSubscriber? = null
@@ -36,12 +38,21 @@ class RNSession(
     WebView.setWebContentsDebuggingEnabled(BuildConfig.DEBUG)
     webView.settings.setJavaScriptEnabled(true)
     webView.addJavascriptInterface(JavaScriptInterface(), "AndroidInterface")
+    setUserAgentString(webView, applicationNameForUserAgent)
     webView.webChromeClient = RNWebChromeClient(reactContext, this@RNSession)
     session.isRunningInAndroidNavigation = false
     session
   }
   val webView: TurboWebView get() = turboSession.webView
   val currentVisit: TurboVisit? get() = turboSession.currentVisit
+
+  private fun setUserAgentString(webView: TurboWebView, applicationNameForUserAgent: String?) {
+    var userAgentString = WebSettings.getDefaultUserAgent(webView.context)
+    if (applicationNameForUserAgent != null) {
+      userAgentString = "$userAgentString $applicationNameForUserAgent"
+    }
+    webView.settings.userAgentString = userAgentString
+  }
 
   internal fun registerVisitableView(newView: SessionSubscriber) {
     visitableView = newView
