+++ packages/turbo/android/src/main/java/com/reactnativeturbowebview/RNSessionManager.kt
@@ -18,8 +18,9 @@ class RNSessionManager(reactContext: ReactApplicationContext) :
     fun findOrCreateSession(
       reactContext: ReactApplicationContext,
       sessionHandle: String,
+      applicationNameForUserAgent: String?
     ): RNSession = sessions.getOrPut(sessionHandle) {
-      RNSession(reactContext, sessionHandle)
+      RNSession(reactContext, sessionHandle, applicationNameForUserAgent)
     }
 
     fun clearSnapshotCaches() {
